requirements.txt
# requirements.txt
requests
python-dotenv
colorama
telethon
urllib3